refactor: upcoming shifts
(cherry picked from commit 1ad4d5f6f57194f5097a127d98aeb0062c398c86)
