+++ frontend/src/data/attendance.js
@@ -19,27 +19,6 @@ export const getTotalDays = (shift) => {
 	return toDate.diff(fromDate, "d") + 1
 }
 
-const getShiftDates = (shift) => {
-	const startDate = dayjs(shift.start_date).format("D MMM")
-	const endDate = shift.end_date ? dayjs(shift.end_date).format("D MMM") : "Ongoing"
-	return startDate == endDate ? startDate : `${startDate} - ${endDate}`
-}
-
-const getTotalShiftDays = (shift) => {
-	if (!shift.end_date) return null
-	const end_date = dayjs(shift.end_date)
-	const start_date = dayjs(shift.start_date)
-	return end_date.diff(start_date, "d") + 1
-}
-
-const getShiftTiming = (shift) => {
-	return (
-		shift.start_time.split(":").slice(0, 2).join(":") +
-		" - " +
-		shift.end_time.split(":").splice(0, 2).join(":")
-	)
-}
-
 export const calendarEvents = createResource({
 	url: "hrms.api.get_attendance_calendar_events",
 	auto: true,
@@ -50,25 +29,5 @@ export const calendarEvents = createResource({
 			from_date: firstOfMonth.value.format("YYYY-MM-DD"),
 			to_date: firstOfMonth.value.endOf("M").format("YYYY-MM-DD"),
 		}
-	},
-})
-
-export const shifts = createResource({
-	url: "hrms.api.get_shifts",
-	auto: true,
-	cache: "hrms:shifts",
-	makeParams() {
-		return {
-			employee: employeeResource.data.name,
-		}
-	},
-	transform: (data) => {
-		return data.map((assignment) => {
-			assignment.is_upcoming = !assignment.end_date || dayjs(assignment.end_date).isAfter(dayjs())
-			assignment.shift_dates = getShiftDates(assignment)
-			assignment.total_shift_days = getTotalShiftDays(assignment)
-			assignment.shift_timing = getShiftTiming(assignment)
-			return assignment
-		})
 	},
 })
